delete the training part and ui part
