--- go-webui.bat
@@ -1,2 +1,0 @@
-runtime\python.exe webui.py zh_CN
-pause
