--- go-webui.ps1
@@ -1,4 +1,0 @@
-$ErrorActionPreference = "SilentlyContinue"
-chcp 65001
-& "$PSScriptRoot\runtime\python.exe" "$PSScriptRoot\webui.py" zh_CN
-pause
